// frontend/web/src/services/API/proserve-wb-packaging-api/models/CreateMandatoryComponentsListRequest.ts
export interface CreateMandatoryComponentsListRequest {
    /**
     * Platform for which the mandatory components list is enforced
     * @type {string}
     * @memberof CreateMandatoryComponentsListRequest
     */
    mandatoryComponentsListPlatform: string;
    /**
     * OS for which the mandatory components list is enforced
     * @type {string}
     * @memberof CreateMandatoryComponentsListRequest
     */
    mandatoryComponentsListOsVersion: string;
    /**
     * Architecture for which the mandatory components list is enforced
     * @type {string}
     * @memberof CreateMandatoryComponentsListRequest
     */
    mandatoryComponentsListArchitecture: string;
    /**
     * Components versions that will be prepended to recipe versions
     * @type {Array<ComponentVersionEntry>}
     * @memberof CreateMandatoryComponentsListRequest
     */
    prependedComponentsVersions: Array<ComponentVersionEntry>;
    /**
     * Components versions that will be appended to recipe versions
     * @type {Array<ComponentVersionEntry>}
     * @memberof CreateMandatoryComponentsListRequest
     */
    appendedComponentsVersions: Array<ComponentVersionEntry>;
}